day3 task/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]
